+++ csgo/visualization/__init__.py
@@ -1,0 +1,1 @@
+SIDE_COLORS = {"ct": "#5d79ae", "t": "#de9b35"}
