+++ tests/test_nav.py
@@ -1,7 +1,7 @@
 import pytest
 import numpy as np
 
-from csgo import NAV
+from csgo.data import NAV
 from csgo.analytics.nav import area_distance, find_closest_area, point_distance, point_in_area, PlaceEncoder
 
 class TestNav:
